Separate Vitest-dependent utils into a different entrypoint
