+++ .changeset/curly-years-sparkle.md
@@ -1,0 +1,5 @@
+---
+'@steiger/toolkit': patch
+---
+
+Separate Vitest-dependent utilities into a different entrypoint to make Vitest a truly optional peer dependency
