+++ packages/steiger-plugin-fsd/src/no-file-segments/index.spec.ts
@@ -1,6 +1,6 @@
 import { expect, it } from 'vitest'
 
-import { compareMessages, joinFromRoot, parseIntoFolder as parseIntoFsdRoot } from '@steiger/toolkit'
+import { compareMessages, joinFromRoot, parseIntoFolder as parseIntoFsdRoot } from '@steiger/toolkit/test'
 import noFileSegments from './index.js'
 
 it('reports no errors on a project with only folder segments', async () => {
