+++ packages/steiger-plugin-fsd/src/public-api/index.spec.ts
@@ -1,7 +1,7 @@
 import { expect, it } from 'vitest'
 
 import publicApi from './index.js'
-import { compareMessages, joinFromRoot, parseIntoFolder as parseIntoFsdRoot } from '@steiger/toolkit'
+import { compareMessages, joinFromRoot, parseIntoFolder as parseIntoFsdRoot } from '@steiger/toolkit/test'
 
 it('reports no errors on a project with all the required public APIs', () => {
   const root = parseIntoFsdRoot(`
