+++ packages/steiger/src/features/choose-root-folder/choose-from-similar.ts
@@ -99,7 +99,7 @@ async function resolveWithCorrections(path: string) {
 if (import.meta.vitest) {
   const { test, expect, vi } = import.meta.vitest
   const { vol } = await import('memfs')
-  const { joinFromRoot } = await import('@steiger/toolkit')
+  const { joinFromRoot } = await import('@steiger/toolkit/test')
 
   vi.mock('node:fs/promises', () => import('memfs').then((memfs) => memfs.fs.promises))
