+++ packages/toolkit/package.json
@@ -12,8 +12,14 @@
     "typecheck": "tsc --noEmit"
   },
   "exports": {
-    "types": "./dist/index.d.ts",
-    "import": "./dist/index.js"
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    },
+    "./test": {
+      "types": "./dist/test.d.ts",
+      "import": "./dist/test.js"
+    }
   },
   "type": "module",
   "license": "MIT",
