+++ packages/toolkit/src/index.ts
@@ -4,5 +4,3 @@ export { findAllRecursively } from './find-all-recursively.js'
 export { enableAllRules, createConfigs } from './create-configs.js'
 export { createPlugin } from './create-plugin.js'
 export type { ConfigObjectOf } from './config-object-of.js'
-
-export { compareMessages, createFsMocks, joinFromRoot, parseIntoFolder } from './prepare-test.js'
