+++ packages/toolkit/tsup.config.ts
@@ -1,10 +1,10 @@
 import { defineConfig } from 'tsup'
 
 export default defineConfig({
-  entry: ['src/index.ts'],
+  entry: ['src/index.ts', 'src/test.ts'],
   format: ['esm'],
   dts: {
-    entry: 'src/index.ts',
+    entry: ['src/index.ts', 'src/test.ts'],
     resolve: true,
   },
   treeshake: true,
